API output for Notifications and Volunteer History
