+++ views/pages/index.ejs
@@ -14,8 +14,8 @@
 			<div class="nav-links" id="nav-links">
 				<a class="nav-link" href="/">Home</a>
 				<a class="nav-link" href="/login">Log in</a>
-				<a class="nav-link" href="/user/notifications">Notifications</a>
-				<a class="nav-link" href="/organizer/volunteer-history">History</a>
+				<a class="nav-link" href="/notifications">Notifications</a>
+				<a class="nav-link" href="/volunteer-history">History</a>
 			</div>
 		</nav>
