+++ views/pages/user/notifications.ejs
@@ -8,19 +8,19 @@
 </head>
 
 <body>
+  <nav>
+    <a href="/">Home</a>
+    <a href="/notifications">Notifications</a>
+  </nav>
 
   <section id="notificationSection">
     <h2>Notifications</h2>
 
     <div id="notificationList">
-      
-      <div class="notification" id="notification1">
-        <!-- Dynamic Javascript Implementation -->
-      </div>
-
+      <!-- Notifications will be dynamically inserted here by notifications.js -->
     </div>
   </section>
 
-  <script src="/js/notifications.js"></script>
+  <script src="/js/notifications.js"></script> <!-- Include your notifications JS file here -->
 </body>
 </html>
